Refactoring utility functions into separate file
