+++ transform.js
@@ -1,6 +1,10 @@
-var _ = require('lodash');
-// var describe = require('jscodeshift-helper').describe;
+'use strict';
 
+const _ = require('lodash');
+const utils = require('./utils');
+
+// leaving below describe in here for future easy troubleshooting
+// var describe = require('jscodeshift-helper').describe;
 module.exports = function (fileInfo, api, options) {
   const j = api.jscodeshift;
   const root = j(fileInfo.source);
@@ -26,13 +30,13 @@ module.exports = function (fileInfo, api, options) {
 
         let methodName = p.node.key.name;
         let methodBlockBody = p.node.body.body;
-        let linenum = PRINT_LINE_NUMBERS ? `${getFunctionStartLineNumber(p)}:` : '';
+        let linenum = PRINT_LINE_NUMBERS ? `${utils.getFunctionStartLineNumber(p)}:` : '';
 
-        let relPathToFile = calculatedRelPath(filepath, relPath);
+        let relPathToFile = utils.calculatedRelPath(filepath, relPath);
 
         // Check for super() call
         if (methodBlockBody.length > 0) {
-          let hasSuperCall = hasSuper(methodBlockBody[0]);
+          let hasSuperCall = utils.hasSuper(methodBlockBody[0]);
           if (hasSuperCall) {
             methodBlockBody.splice(1, 0,
               j.expressionStatement(
@@ -44,7 +48,7 @@ module.exports = function (fileInfo, api, options) {
             )
           } else {
             if (shouldLogParams) {
-              let paramsList = buildParamLoggingList(p.node.params, relPathToFile, linenum, '');
+              let paramsList = utils.buildParamLoggingList(p.node.params, relPathToFile, linenum, '');
 
               // Reverse the paramList, as we're unshifted from last parameter to first
               let reversedParamList = paramsList.reverse();
@@ -80,12 +84,12 @@ module.exports = function (fileInfo, api, options) {
         let functionBlockBody = p.node.body.body;
         let functionName = p.node.id.name;
 
-        const paramString = buildAnonymousParamsList(p.node.params);
-        let relPathToFile = calculatedRelPath(filepath, relPath);
-        let linenum = PRINT_LINE_NUMBERS ? `${getFunctionStartLineNumber(p)}:` : '';
+        const paramString = utils.buildAnonymousParamsList(p.node.params);
+        let relPathToFile = utils.calculatedRelPath(filepath, relPath);
+        let linenum = PRINT_LINE_NUMBERS ? `${utils.getFunctionStartLineNumber(p)}:` : '';
 
         if (shouldLogParams) {
-          let paramsList = buildParamLoggingList(p.node.params, relPathToFile, linenum, functionName);
+          let paramsList = utils.buildParamLoggingList(p.node.params, relPathToFile, linenum, functionName);
 
           // Reverse the paramList, as we're unshifted from last parameter to first
           let reversedParamList = paramsList.reverse();
@@ -118,12 +122,12 @@ module.exports = function (fileInfo, api, options) {
       .forEach(p => {
         let functionBlockBody = p.node.body.body;
 
-        const paramString = buildAnonymousParamsList(p.node.params);
-        let relPathToFile = calculatedRelPath(filepath, relPath);
-        let linenum = PRINT_LINE_NUMBERS ? `${getFunctionStartLineNumber(p)}:` : '';
+        const paramString = utils.buildAnonymousParamsList(p.node.params);
+        let relPathToFile = utils.calculatedRelPath(filepath, relPath);
+        let linenum = PRINT_LINE_NUMBERS ? `${utils.getFunctionStartLineNumber(p)}:` : '';
 
         if (shouldLogParams) {
-          let paramsList = buildParamLoggingList(p.node.params, relPathToFile, linenum, '');
+          let paramsList = utils.buildParamLoggingList(p.node.params, relPathToFile, linenum, '');
 
           // Reverse the paramList, as we're unshifted from last parameter to first
           let reversedParamList = paramsList.reverse();
@@ -154,11 +158,11 @@ module.exports = function (fileInfo, api, options) {
   const addLoggingToArrowFunctions = (path, filepath, shouldLogParams) => {
     path.find(j.ArrowFunctionExpression)
       .forEach(p => {
-        const paramString = buildAnonymousParamsList(p.node.params);
+        const paramString = utils.buildAnonymousParamsList(p.node.params);
 
         let blockStatementBody = _.get(p, 'node.body.body', false);
-        let relPathToFile = calculatedRelPath(filepath, relPath);
-        let linenum = PRINT_LINE_NUMBERS ? `${getFunctionStartLineNumber(p)}:` : '';
+        let relPathToFile = utils.calculatedRelPath(filepath, relPath);
+        let linenum = PRINT_LINE_NUMBERS ? `${utils.getFunctionStartLineNumber(p)}:` : '';
 
         if (blockStatementBody) {
 
@@ -169,7 +173,7 @@ module.exports = function (fileInfo, api, options) {
 
           if (Array.isArray(blockStatementBody)) {
             if (shouldLogParams) {
-              let paramsList = buildParamLoggingList(p.node.params, relPathToFile, linenum, '');
+              let paramsList = utils.buildParamLoggingList(p.node.params, relPathToFile, linenum, '');
 
               // Reverse the paramList, as we're unshifted from last parameter to first
               let reversedParamList = paramsList.reverse();
@@ -244,7 +248,7 @@ module.exports = function (fileInfo, api, options) {
           return;
         }
 
-        let relPathToFile = calculatedRelPath(filepath, relPath);
+        let relPathToFile = utils.calculatedRelPath(filepath, relPath);
         let linenum = PRINT_LINE_NUMBERS ? `${p.node.loc.start.line}` : '';
 
         p.insertBefore(
@@ -275,7 +279,7 @@ module.exports = function (fileInfo, api, options) {
           return;
         }
 
-        let relPathToFile = calculatedRelPath(filepath, relPath);
+        let relPathToFile = utils.calculatedRelPath(filepath, relPath);
         let linenum = PRINT_LINE_NUMBERS ? `${p.node.loc.start.line}` : '';
 
         p.insertBefore(
@@ -336,7 +340,7 @@ module.exports = function (fileInfo, api, options) {
         //
         if (!pipeStatementFound) {
           pipeStatementFound = true;
-          let tapImportExists = findTapImport(path);
+          let tapImportExists = utils.findTapImport(path);
 
           if (tapImportExists === false) {
             // If the tap import hasn't been added, then go ahead and add it
@@ -363,7 +367,7 @@ module.exports = function (fileInfo, api, options) {
 
         for (let i = 0; i < totalPipeParametersPlusFinished; i++) {
           let lineNumber = p.node.loc.start.line;
-          let arrowFunc = printRxjsPipeStageLogFunction(totalPipeParameters, i, lineNumber, filepath);
+          let arrowFunc = utils.printRxjsPipeStageLogFunction(totalPipeParameters, i, lineNumber, filepath, relPath);
           let tapExpressionStatement = j.callExpression(j.identifier('tap'), [arrowFunc]);
           newArgArray.push(tapExpressionStatement);
           newArgArray.push(p.node.arguments[i]);
@@ -371,171 +375,6 @@ module.exports = function (fileInfo, api, options) {
 
         p.node.arguments = newArgArray;
       })
-  }
-
-  /** @function
-   * @name buildAnonymousParamsList
-   * @param A list of parameter nodes
-   * @returns {string} A string representation of function parameters
-   */
-  const buildAnonymousParamsList = (paramNodes) => {
-    let paramString = '(';
-
-    for (let index = 0; index < paramNodes.length; index++) {
-      paramString = paramString + paramNodes[index].name;
-
-      if (index !== (paramNodes.length - 1)) {
-        paramString = paramString + ', ';
-      }
-    }
-
-    paramString = paramString + ')';
-    return paramString;
-  }
-
-  /** @function
-   * @name @buildParamLoggingList
-   * @param { Object } paramNodes A list of parameter nodes
-   * @returns { Object } An array of nodes representing a console.log() node for each paraemter
-   */
-  const buildParamLoggingList = (paramNodes, relPathToFile, linenum, functionName) => {
-    const returnExpressionNodes = [];
-    for (let index = 0; index < paramNodes.length; index++) {
-      let currentNode = paramNodes[index];
-
-      // This part is to deal with situations where we're handed a TypeScript parameter with a type annotion.
-      // If this is this case, then there's an additional "left" parent attribute that needs to be accessed
-      // e.g.
-      let currentNodeName = _.get(currentNode, 'left.name', currentNode.name);
-
-      let announcement = `${LIA_PREFIX}\t${relPathToFile}:${linenum}${functionName}:param ${currentNodeName} value:${LIA_SUFFIX} \n\t\t\t\t\t\t\t\t\t\t`;
-
-      let quasis = [
-        j.templateElement({ cooked: announcement, raw: announcement }, true),
-      ];
-
-      let stringifyCallExpression = j.callExpression(j.identifier('JSON.stringify'), [j.identifier(currentNodeName)]);
-      let logTemplateLiteral = j.templateLiteral(quasis, [stringifyCallExpression]);
-      let consoleCallExpression = j.callExpression(j.identifier('console.log'), [logTemplateLiteral]);
-      let expressionStatement = j.expressionStatement(consoleCallExpression);
-
-      returnExpressionNodes.push(expressionStatement);
-    }
-
-    return returnExpressionNodes;
-  }
-
-  /** @function
-   * @name printRxjsPipeStageLogFunction
-   * @param totalPipeParameters The total number of page parameters
-   * @param pipeStageIndex The number value
-   * @returns {Object} A jscodeshift ArrowFunction object prints the pipe state and value
-   *
-   *
-   * The end result of this function should a jscodeshift ArrowFunction object
-   * that when rendered looks like the following
-
-        x => {
-          console.log(`Stage 1 value for pipe at line 1 is: ${x}\n`);
-        }
-   */
-  const printRxjsPipeStageLogFunction = (totalPipeParameters, pipeStageIndex, pipeStartLineNumber, filepath) => {
-    let announcement = '';
-    let relPathToFile = calculatedRelPath(filepath, relPath);
-
-    if (totalPipeParameters === pipeStageIndex) {
-      announcement = `${LIA_PREFIX}Final value for rxjs pipe starting at line ${pipeStartLineNumber} in ${relPathToFile} is:\n`;
-    } else {
-      announcement = `${LIA_PREFIX}Stage ${pipeStageIndex} value for rxjs pipe starting at line ${pipeStartLineNumber} in ${relPathToFile} is:\n`;
-    }
-
-    let quasis = [
-      j.templateElement({ cooked: announcement, raw: announcement }, true),
-      j.templateElement({ cooked: '\\n', raw: '\\n' }, true)
-    ];
-
-    let expressions = [j.identifier('x')];
-    let logTemplateLiteral = j.templateLiteral(quasis, expressions);
-
-    let expressionStatement = j.expressionStatement(j.callExpression(j.identifier('console.log'), [logTemplateLiteral]))
-    let arrowFunction = j.arrowFunctionExpression([j.identifier('x')],
-      j.blockStatement([expressionStatement]));
-    return arrowFunction;
-  }
-
-
-  /** @function
-   * @name hasSuper
-   * @param A jscodeshift path
-   * @returns {boolean} True/false value as to whether there's a super call in the constructor
-   *
-   * Checks whether there's a super call that will require the log statement to be put immediately after it
-  */
-  const hasSuper = (p) => {
-    let returnSuper = false;
-
-    if (_.get(p, 'expression.type', false)) {
-      let calleeType = _.get(p, 'expression.callee.type', '');
-      if (calleeType === 'Super') {
-        returnSuper = true;
-      }
-    }
-
-    return returnSuper;
-  }
-
-  /** @function
-   * @name calculatedRelPath
-   * @param fullpath String value with the absolute path
-   * @param relpath A String value with the relative path
-   * @returns {string} A shortened version of the absolute path
-   *
-   * Whittles down the absolute path into something more managable
-   * to look at in console.log() statements
-   */
-  const calculatedRelPath = (fullpath, relpath) => {
-    let foundIndex = fullpath.search(relpath);
-    let relPathToFile = fullpath.substring(foundIndex);
-    return relPathToFile;
-  }
-
-  /** @function
-   * @name getFunctionStartLineNumber
-   * @param path A jscodeshift path
-   * @returns {integer} The functions starting line number
-   *
-   * Convenience function that returns the starting line number from a
-   * jscodeshift path
-  */
-  const getFunctionStartLineNumber = (path) => {
-    return path.node.body.loc.start.line;
-  }
-
-  /** @function
-   * @name findTapImport
-   * @param path A jscodeshift node
-   * @returns {boolean} Flag whether there's an existing rxjs tap operator
-   *
-   * This function is used as a filter function on ImportDeclaration nodes
-   * to find out whether the ImportDeclaration imports the rxjs tap operator
-  */
-  const findTapImport = (path) => {
-    let tapImportExists = false;
-    path.find(j.ImportDeclaration)
-      .forEach(p => {
-        let importSource = _.get(p, 'node.source.value');
-        if ((importSource === 'rxjs/operators') ||
-          (importSource === 'rxjs/internal/operators')) {
-          _.find(p.node.specifiers, x => {
-            let importName = _.get(x, 'imported.name', '');
-            if (importName === 'tap') {
-              tapImportExists = true;
-            }
-          })
-        }
-      });
-
-    return tapImportExists;
   }
 
   // MAIN SECTION
